Improve precision of tagtrail_ocr by identifying sheets conservatively
The worst kind of error which can happen during tagtrail_ocr is
confounding the name and sheet number of a scanned sheet. This kind of
error is difficult to recognize and correct during tagtrail_sanitize.

To increase precision of tagtrail_ocr on nameBox and sheetNumberBox,
they are validated based on our known input sheets now and their
confidence set to 0 (user has to verify them during tagtrail_sanitize,
recall is reduced) if there is too little evidence that they have been
recognized correctly.