Disable unit and price editing - they can be loaded from existing sheet
As the price can differ between sheets and products.csv (tagtrail_gen
does not discard sheets for minor price changes), the price of a product
sheet should not be entered manually any more, but loaded from the input
sheet.

If all works normally, this information is guaranteed to be available as
each scanned sheet has an input sheet counterpart which is already
generated with price and unit printed on it.

Through user errors, it is possible to loose an input image and scan it
again later (most likely because it was full and should have been
removed during last accounting, but the user forgot to do so - or
because somebody tampered with the files in between accountings).
To offer a best effort workaround for these scenarios, a dialog giving
the user the necessary means and advice to recreate a lost input sheet
has been added.