tagtrail_sanitize can switch between duplicates
If the currently loaded sheet to be sanitized cannot be stored because
a sheet of the same name already exists, the user now has the
possibility to switch to the existing sheet.

This is necessary as one of the two sheets must have been wrongly
identified, but it doesn't need to be the currently loaded sheet.