Removed possibility for user to reassign categories in tagtrail_gen
tagtrail_gen automatically assigns categories to sheets (active,
inactive, to be replaced, obsolete). The user should not be able to mess
around with this system, as it is hard to understand the effects it
would have if e.g. a sheet is inactivated that should be active.

As an example, tagtrail_gen might need to make room for 22 tags of a
given product and foresees sheet 1 to be activated, while sheet 2 should
be inactive. For the user, it is not obvious if he can deactivate sheet
1 and replace it with sheet 1, as they might have a different number of
open tags left.