Avoid filling black pixels when rotating relevant tag area
When a tag was sticked close to the border of its box, the relevant
contours could be too close to the border of the box. Therefore, upon
rotating the relevant area for OCR, black pixels were filled in. This
patch fixes the problem by taking a larger region around the box into
account for rotation and cropping.