Checking validity of Postfinance Export file
Postfinance currently has a bug in its export process. It includes
bookings out of the selected range. To remedy the situation and as part
of a general data sanity check, tagtrail_bankimport has been changed to
check each booking if it is in the expected time range when loading the
export file.